SI-164: additional work on impala build

+++ sh/clean.sh
@@ -12,5 +12,11 @@ clean()
 {
     if (($HELP)) ; then clean_help ; return ; fi
 
+    if [ -z "$WORKSPACE" ] ; then echo "clean requires WORKSPACE" ; exit 1 ; fi
+
+    for mockdir in ${WORKSPACE}/build-*/mock ; do
+	mock --configdir=$mockdir -r mock --clean
+    done
+
     rm -rf ${WORKSPACE}/build-*
 }
